+++ vulp/__init__.py
@@ -6,4 +6,4 @@
 
 """Real-time motion control for Python."""
 
-__version__ = "2.0.0"
+__version__ = "2.1.0"