+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Base class ``VulpException`` for all exceptions
 - Check that msgpack is not running the pure-Python implementation
-- Python: ``vulp.utils.exceptions`` submodule
+- New ``PerformanceIssue`` exception
 - Take feedforward torque commands into account in all actuation interfaces
 
 ## [2.0.0] - 2023-11-15