+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [2.1.0] - 2024-01-03
+
 ### Added
 
 - Base class ``VulpException`` for all exceptions